+++ lib/http_cookie.mli
@@ -21,34 +21,41 @@
     The standard implemented by the library is
     {{:https://tools.ietf.org/html/rfc6265} RFC 6265}. *)
 
-(** {1 Types} *)
-
-(** Represents 'Same-site' cookie attribute. See
-    https://tools.ietf.org/html/draft-ietf-httpbis-cookie-same-site-00. *)
-module Same_site : sig
-  type t = Default | None | Lax | Strict
-
-  val equal : t -> t -> bool
-  val compare : t -> t -> int
-  val to_string : t -> string
-end
-
-(** Represents a cookie name-value in [Cookie] request header or a set of cookie
-    attributes in [Set-Cookie] response header. *)
+(** A HTTP cookie. *)
 type t
 
-type date_time =
-  { year: int  (** Four digit year value, e.g. 2020, 2019, 1980 etc. *)
-  ; month: int  (** Begins from 0, i.e. January = 0. *)
-  ; weekday: [`Sun | `Mon | `Tue | `Wed | `Thu | `Fri | `Sat]
-  ; day_of_month: int  (** Day of the month value from 1 - 31. *)
-  ; hour: int  (** 24 hour value from 0-23 *)
-  ; minutes: int  (** Minutes value from 0 - 59*)
-  ; seconds: int  (** Seconds value from 0 - 60 *) }
+(** Cookie expiry datetime value. *)
+and date_time
 
-exception Cookie of string
+(** 'Same-site' cookie attribute. See
+    https://tools.ietf.org/html/draft-ietf-httpbis-cookie-same-site-00. *)
+and same_site = [`None | `Lax | `Strict]
 
-(** {1 Constructors} *)
+(** {1 Http_cookie} *)
+
+val date_time :
+     year:int
+  -> month:
+       [ `Jan
+       | `Feb
+       | `Mar
+       | `Apr
+       | `May
+       | `Jun
+       | `Jul
+       | `Aug
+       | `Sep
+       | `Oct
+       | `Nov
+       | `Dec ]
+  -> weekday:[`Sun | `Mon | `Tue | `Wed | `Thu | `Fri | `Sat]
+  -> day_of_month:int
+  -> hour:int (** 24 hour format *)
+  -> minutes:int
+  -> seconds:int
+  -> (date_time, string) result
+(** [date_time] is [Ok dt] if all of the given parameters are valid for creating
+    {!type:date_time} value, otherwise [Error err] is denotes the error. *)
 
 val create :
      ?path:string
@@ -57,32 +64,40 @@ val create :
   -> ?max_age:int
   -> ?secure:bool
   -> ?http_only:bool
-  -> ?same_site:Same_site.t
+  -> ?same_site:same_site
   -> ?extension:string
+  -> name:string
   -> string
-  -> value:string
-  -> t
-(** [create ~path ~domain ~expires ~max_age ~secure ~http_only ~same_site ~extension name
-    ~value]
-    returns a cookie instance {!type:t} with cookie name [name] and value
-    [value] along with the given attributes.
+  -> (t, string) result
+(** [create ~path ~domain ~expires ~max_age ~secure ~http_only ~same_site ~extension ~name
+    value]
+    is [Ok cookie] if all of the given parameters are valid cookie attribute
+    values. Otherwise it is [Error error] where [error] is the description of
+    the error. *)
 
-    @raise Cookie if any of the cookie attributes fail validation. *)
-
-val of_cookie_header : string -> t list
-(** [of_cookie_header s] parses [s] - a string value which represents HTTP
+val of_cookie : string -> (t list, string) result
+(** [of_cookie header] parses [header] - a string value which represents HTTP
     [Cookie] header value as defined in
-    {:https://tools.ietf.org/html/rfc6265#section-4.2} and returns a list of
-    [Cookie]s.
+    {:https://tools.ietf.org/html/rfc6265#section-4.2}. It returns a list of
+    [Cookie]s if it is able to successfully parse [s], otherwise it returns
+    [Error err].
 
     {4 Examples}
 
-    This returns two cookies with cookie names [SID] and [lang-en].
+    This returns two cookies with cookie names [SID] and [lang].
 
-    {[ Http_cookie.of_cookie_header "SID=31d4d96e407aad42; lang=en-US" ]} *)
+    {[ Http_cookie.of_cookie "SID=31d4d96e407aad42; lang=en-US" ]} *)
 
-val to_set_cookie_header_value : t -> string
-(** [to_set_header c] serializes cookie [c] into a string which can be encoded
+val to_cookie : t -> string
+(** [to_cookie c] serializes [c] into a string which can be encoded as value for
+    HTTP [Cookie] header.
+
+    Example of a string returned by the function.
+
+    {v SID=31d4d96e407aad42 v} *)
+
+val to_set_cookie : t -> string
+(** [to_set_cookie c] serializes cookie [c] into a string which can be encoded
     as value for HTTP [Set-Cookie] header.
 
     The datetime format for [expires] attribute is specified in
@@ -94,15 +109,7 @@ val to_set_cookie_header_value : t -> string
 SID=31d4d96e407aad42; Path=/; Secure; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT
     v} *)
 
-val to_cookie_header_value : t -> string
-(** [to_cookie_header c] serializes [c] into a string which can be encoded as
-    value for HTTP [Cookie] header.
-
-    Example of a string returned by the function.
-
-    {v SID=31d4d96e407aad42 v} *)
-
-(** {1 Query Cookie Attributes}
+(** {1 Cookie Attributes}
 
     Cookie attributes are defined precisely at
     {{:https://tools.ietf.org/html/rfc6265#section-4.1.1} RFC 6262} *)
@@ -148,7 +155,7 @@ val http_only : t -> bool option
 
     See {{:https://tools.ietf.org/html/rfc6265#section-4.1.2.6} http-only} *)
 
-val same_site : t -> Same_site.t option
+val same_site : t -> same_site option
 (** [same_site t] returns a same_site attribute.
 
     See {{:https://tools.ietf.org/html/draft-ietf-httpbis-cookie-same-site-00}
@@ -166,15 +173,23 @@ val compare : t -> t -> int
     if [c1] is greater than [c2] and a negative integer if [c1] is less than
     [c2] *)
 
+val compare_date_time : date_time -> date_time -> int
+
 (** {1 Updates} *)
 
-val update_value : string -> t -> t
-val update_name : string -> t -> t
-val update_path : string option -> t -> t
-val update_domain : string option -> t -> t
+val update_value : string -> t -> (t, string) result
+val update_name : string -> t -> (t, string) result
+val update_path : string option -> t -> (t, string) result
+val update_domain : string option -> t -> (t, string) result
 val update_expires : date_time option -> t -> t
-val update_max_age : int option -> t -> t
+val update_max_age : int option -> t -> (t, string) result
 val update_secure : bool option -> t -> t
 val update_http_only : bool option -> t -> t
-val update_same_site : Same_site.t option -> t -> t
-val update_extension : string option -> t -> t
+val update_same_site : same_site option -> t -> t
+val update_extension : string option -> t -> (t, string) result
+
+(** {1 Pretty Printers} *)
+
+val pp : Format.formatter -> t -> unit
+val pp_date_time : Format.formatter -> date_time -> unit
+val pp_same_site : Format.formatter -> same_site -> unit